feat: change routes and update content
